requirements.txt
discord.py-self
pycryptodomex
pypiwin32
toasted
pystray
Pillow
PyQt6-Fluent-Widgets